+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add function reInit() widget
 - Fixes bug 'undefined' on localstorage
 - Fixes bug refresh list conversations
+- Add Attachment in Attribute msg (buttons)
 
 ### 1.019
 - Load variables from server by projectId